:art: improve structure: mainly reorganize stores / hooks & utils (and some clean elements)
