+++ src/renderer/styles/global.scss
@@ -1,6 +1,5 @@
 @use "./font";
 @use "./column";
-// TODO: Mettre scss files dans un dossier styles pour gérer les mixins et variables. (à check ensemble @lsagetlethias)
 @import "./colors";
 
 * {
