+++ src/renderer/views/dashboard/DashboardInformationsMail.tsx
@@ -3,8 +3,8 @@ import type { ComputedDatum } from "@nivo/circle-packing";
 import type { FC } from "react";
 import React from "react";
 
-import type { MailViewerObject } from "../../utils/pst-extractor";
-import { sanitizeMailDate } from "../../utils/pst-viewer";
+import { sanitizeMailDate } from "../../utils/dashboard-viewer";
+import type { MailViewerObject } from "../../utils/dashboard-viewer-dym";
 
 export const DashboardInformationsMail: FC<{
     mainInfos: ComputedDatum<MailViewerObject<string>>;
