+++ src/renderer/views/start-screen/StartScreen.module.scss
@@ -6,10 +6,12 @@
   text-align: center;
 }
 
-/* TODO: handle properly the logo */
+
 .logo {
-  height: 5em;
-  margin-top: 1em;
+  position: absolute;
+  width: 10rem;
+  left: 2rem;
+  top: 2rem;
 }
 
 .startscreen__progress {
