+++ src/renderer/views/start-screen/StartScreen.tsx
@@ -3,7 +3,7 @@ import React, { useCallback, useEffect } from "react";
 import { StaticImage } from "../../components/common/staticImage/StaticImage";
 import { Dropzone } from "../../components/dropzone/Dropzone";
 import { useRouteContext } from "../../context/RouterContext";
-import { usePstExtractor } from "../../hooks/usePSTExtractor";
+import { usePstExtractor } from "../../hooks/usePstExtractor";
 import { usePstStore } from "../../store/PSTStore";
 import { ACCEPTED_EXTENSION } from "../../utils/constants";
 import style from "./StartScreen.module.scss";
