+++ static/locales/en-GB/translation.json
@@ -2,6 +2,39 @@
   "app.quit.detail": "Data not saved!",
   "app.quit.leave": "Leave?",
   "app.quit.title": "Quit app",
+  "dashboard.impact.cardTitle": "",
+  "dashboard.informations.attachedCount": "",
+  "dashboard.informations.bcc": "",
+  "dashboard.informations.cardTitle": "",
+  "dashboard.informations.cc": "",
+  "dashboard.informations.emptyInfos": "",
+  "dashboard.informations.folder": "",
+  "dashboard.informations.from": "",
+  "dashboard.informations.mail": "",
+  "dashboard.informations.mailCount": "",
+  "dashboard.informations.mailFocus": "",
+  "dashboard.informations.percentage": "",
+  "dashboard.informations.receivedDate": "",
+  "dashboard.informations.sentDate": "",
+  "dashboard.informations.title": "",
+  "dashboard.informations.to": "",
+  "dashboard.informations.type": "",
+  "dashboard.informations.viewer.breadcrumb.domain": "",
+  "dashboard.informations.viewer.breadcrumb.level": "",
+  "dashboard.informations.viewer.cardTitle": "",
+  "dashboard.recap.cardTitle": "",
+  "dashboard.recap.contacts": "",
+  "dashboard.recap.deletedMessages": "",
+  "dashboard.recap.mails": "",
+  "dashboard.recap.pc": "",
+  "dashboard.recap.receivedMessages": "",
+  "dashboard.recap.selectFolder.choosePlaceholder": "",
+  "dashboard.recap.selectFolder.selector": "",
+  "dashboard.recap.selectFolder.validate": "",
+  "dashboard.recap.sentMessages": "",
+  "dashboard.viewer.breadcrumb.domain": "",
+  "dashboard.viewer.breadcrumb.level": "",
+  "dashboard.viewer.cardTitle": "",
   "dropzone.click-or-drag": "Click or drag'n'drop your files here",
   "dropzone.download-here": "Download your file here",
   "dropzone.drop": "Drop to upload files",
@@ -26,5 +59,15 @@
   "exporter.table.subject": "Subject of the mail",
   "exporter.table.to.email": "Email addresses of recipient(s) A",
   "exporter.table.to.name": "Name of recipient(s) A",
-  "hello": "hello"
+  "hello": "hello",
+  "loader.loading": "Loading...",
+  "startscreen.importInfo.attachedCount": "Attached count",
+  "startscreen.importInfo.emails": "Emails",
+  "startscreen.importInfo.folders": "Folders",
+  "startscreen.importInfo.seconds": "{{seconds}}sec",
+  "startscreen.importInfo.totalFiles": "Total",
+  "startscreen.importInfo.totalTime": "Total time (in seconds)",
+  "tag.markedTags.delete": "To delete",
+  "tag.markedTags.keep": "To conserve",
+  "tag.markedTags.untag": "Unmarked"
 }
