+++ static/locales/fr-FR/translation.json
@@ -2,6 +2,39 @@
   "app.quit.detail": "Données non sauvegardées !",
   "app.quit.leave": "Quitter ?",
   "app.quit.title": "Quitter l'application",
+  "dashboard.impact.cardTitle": "Impact",
+  "dashboard.informations.attachedCount": "Nombre de PJ",
+  "dashboard.informations.bcc": "Bcc",
+  "dashboard.informations.cardTitle": "Informations",
+  "dashboard.informations.cc": "Cc",
+  "dashboard.informations.emptyInfos": "Passer le curseur sur le visualiseur pour afficher des informations",
+  "dashboard.informations.folder": "fichiers",
+  "dashboard.informations.from": "Expéditeur",
+  "dashboard.informations.mail": "mail",
+  "dashboard.informations.mailCount": "Nombre de mails",
+  "dashboard.informations.mailFocus": "Mail (cliquer pour focus)",
+  "dashboard.informations.percentage": "Représentation (en %)",
+  "dashboard.informations.receivedDate": "Date de réception",
+  "dashboard.informations.sentDate": "Date d'envoi",
+  "dashboard.informations.title": "Titre",
+  "dashboard.informations.to": "Destinataire(s)",
+  "dashboard.informations.type": "Type",
+  "dashboard.informations.viewer.breadcrumb.domain": "",
+  "dashboard.informations.viewer.breadcrumb.level": "",
+  "dashboard.informations.viewer.cardTitle": "",
+  "dashboard.recap.cardTitle": "Synthèse",
+  "dashboard.recap.contacts": "Contacts",
+  "dashboard.recap.deletedMessages": "Messages supprimés",
+  "dashboard.recap.mails": "Mails",
+  "dashboard.recap.pc": "pj",
+  "dashboard.recap.receivedMessages": "Messages reçus",
+  "dashboard.recap.selectFolder.choosePlaceholder": "Choisissez votre dossier de messages supprimés",
+  "dashboard.recap.selectFolder.selector": "Pour afficher la synthèse de votre PST, sélectionnez votre dossier de mails supprimés",
+  "dashboard.recap.selectFolder.validate": "Valider mon choix",
+  "dashboard.recap.sentMessages": "Messages envoyés",
+  "dashboard.viewer.breadcrumb.domain": "Domaine",
+  "dashboard.viewer.breadcrumb.level": "Niveau",
+  "dashboard.viewer.cardTitle": "Visualisation",
   "dropzone.click-or-drag": "Cliquez ou glissez-déposez vos fichiers ici",
   "dropzone.download-here": "Téléchargez votre fichier ici",
   "dropzone.drop": "Lâchez pour téléverser les fichiers",
@@ -26,5 +59,15 @@
   "exporter.table.subject": "Objet du mail",
   "exporter.table.to.email": "Adresses mails du ou des destinataire A",
   "exporter.table.to.name": "Noms du ou des destinataire A",
-  "hello": "bonjour"
+  "hello": "bonjour",
+  "loader.loading": "Chargement...",
+  "startscreen.importInfo.attachedCount": "Pièces jointes",
+  "startscreen.importInfo.emails": "Emails",
+  "startscreen.importInfo.folders": "Dossiers",
+  "startscreen.importInfo.seconds": "{{seconds}}s",
+  "startscreen.importInfo.totalFiles": "Total",
+  "startscreen.importInfo.totalTime": "Temps total (en secondes)",
+  "tag.markedTags.delete": "À supprimer",
+  "tag.markedTags.keep": "À conserver",
+  "tag.markedTags.untag": "Non marqué"
 }
